Adding a simple last seen highlight
Adding a feature for higlighting the last video to help with knowing what was watched. This only works for when you are at the lowest part of the directory. If you navigate up it gets lost. This is not ideal, but for simple usecases it works.

Next would probably be to see if it can be tracked within the records sake to know if the item has been observed before so that it can be kept in memory.